+++ src/main/java/com/powsybl/openloadflow/dc/DcIncrementalPhaseControlOuterLoop.java
@@ -12,20 +12,14 @@ import com.powsybl.openloadflow.lf.outerloop.IncrementalContextData;
 import com.powsybl.openloadflow.dc.equations.DcEquationType;
 import com.powsybl.openloadflow.dc.equations.DcVariableType;
 import com.powsybl.openloadflow.equations.EquationSystem;
-import com.powsybl.openloadflow.equations.EquationTerm;
 import com.powsybl.openloadflow.equations.JacobianMatrix;
 import com.powsybl.openloadflow.lf.outerloop.AbstractIncrementalPhaseControlOuterLoop;
 import com.powsybl.openloadflow.lf.outerloop.OuterLoopStatus;
 import com.powsybl.openloadflow.network.*;
-import com.powsybl.openloadflow.util.PerUnit;
-import org.apache.commons.lang3.Range;
-import org.apache.commons.lang3.mutable.MutableBoolean;
-import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -33,44 +27,23 @@ import java.util.Objects;
 public class DcIncrementalPhaseControlOuterLoop
         extends AbstractIncrementalPhaseControlOuterLoop<DcVariableType, DcEquationType, DcLoadFlowParameters, DcLoadFlowContext, DcOuterLoopContext> {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(DcIncrementalPhaseControlOuterLoop.class);
+    public DcIncrementalPhaseControlOuterLoop() {
+        LOGGER = LoggerFactory.getLogger(DcIncrementalPhaseControlOuterLoop.class);
+    }
 
     @Override
     public String getType() {
         return "DC Incremental phase control";
     }
 
-    public static class DcSensitivityContext {
-
-        private final List<LfBranch> controllerBranches;
-
-        private final EquationSystem<DcVariableType, DcEquationType> equationSystem;
-
-        private final JacobianMatrix<DcVariableType, DcEquationType> jacobianMatrix;
-
-        private final int[] controllerBranchIndex;
-
-        private DenseMatrix sensitivities;
-
-        public DcSensitivityContext(LfNetwork network, List<LfBranch> controllerBranches,
-                                  EquationSystem<DcVariableType, DcEquationType> equationSystem,
-                                  JacobianMatrix<DcVariableType, DcEquationType> jacobianMatrix) {
-            this.controllerBranches = Objects.requireNonNull(controllerBranches);
-            this.equationSystem = Objects.requireNonNull(equationSystem);
-            this.jacobianMatrix = Objects.requireNonNull(jacobianMatrix);
-            controllerBranchIndex = LfBranch.createIndex(network, controllerBranches);
+    public static class DcSensitivityContext extends AbstractSensitivityContext<DcVariableType, DcEquationType> {
+        public DcSensitivityContext(LfNetwork network, List<LfBranch> controllerBranches, EquationSystem<DcVariableType, DcEquationType> equationSystem, JacobianMatrix<DcVariableType, DcEquationType> jacobianMatrix) {
+            super(network, controllerBranches, equationSystem, jacobianMatrix);
         }
 
-        private DenseMatrix getSensitivities() {
-            if (sensitivities == null) {
-                sensitivities = calculateSensitivityValues(controllerBranches, controllerBranchIndex, equationSystem, jacobianMatrix);
-            }
-            return sensitivities;
-        }
-
-        private static DenseMatrix calculateSensitivityValues(List<LfBranch> controllerBranches, int[] controllerBranchIndex,
-                                                              EquationSystem<DcVariableType, DcEquationType> equationSystem,
-                                                              JacobianMatrix<DcVariableType, DcEquationType> jacobianMatrix) {
+        public DenseMatrix calculateSensitivityValues(List<LfBranch> controllerBranches, int[] controllerBranchIndex,
+                                                      EquationSystem<DcVariableType, DcEquationType> equationSystem,
+                                                      JacobianMatrix<DcVariableType, DcEquationType> jacobianMatrix) {
             DenseMatrix rhs = new DenseMatrix(equationSystem.getIndex().getSortedEquationsToSolve().size(), controllerBranches.size());
             for (LfBranch controllerBranch : controllerBranches) {
                 equationSystem.getEquation(controllerBranch.getNum(), DcEquationType.BRANCH_TARGET_ALPHA1)
@@ -79,64 +52,6 @@ public class DcIncrementalPhaseControlOuterLoop
             jacobianMatrix.solveTransposed(rhs);
             return rhs;
         }
-
-        @SuppressWarnings("unchecked")
-        private static EquationTerm<DcVariableType, DcEquationType> getP1(LfBranch controlledBranch) {
-            return (EquationTerm<DcVariableType, DcEquationType>) controlledBranch.getP1();
-        }
-
-        @SuppressWarnings("unchecked")
-        private static EquationTerm<DcVariableType, DcEquationType> getP2(LfBranch controlledBranch) {
-            return (EquationTerm<DcVariableType, DcEquationType>) controlledBranch.getP2();
-        }
-
-        double calculateSensitivityFromA2S(LfBranch controllerBranch, EquationTerm<DcVariableType, DcEquationType> s) {
-            return s.calculateSensi(getSensitivities(), controllerBranchIndex[controllerBranch.getNum()]);
-        }
-
-        double calculateSensitivityFromA2P(LfBranch controllerBranch, LfBranch controlledBranch, ControlledSide controlledSide) {
-            var p = controlledSide == ControlledSide.ONE ? getP1(controlledBranch) : getP2(controlledBranch);
-            return calculateSensitivityFromA2S(controllerBranch, p);
-        }
-    }
-
-    private static boolean checkActivePowerControlPhaseControls(DcSensitivityContext sensitivityContext, IncrementalContextData contextData,
-                                                                List<TransformerPhaseControl> activePowerControlPhaseControls) {
-        MutableBoolean updated = new MutableBoolean(false);
-
-        for (TransformerPhaseControl phaseControl : activePowerControlPhaseControls) {
-            LfBranch controllerBranch = phaseControl.getControllerBranch();
-            LfBranch controlledBranch = phaseControl.getControlledBranch();
-            var p = phaseControl.getControlledSide() == ControlledSide.ONE
-                    ? controlledBranch.getP1() : controlledBranch.getP2();
-            double pValue = p.eval();
-            double halfTargetDeadband = getHalfTargetDeadband(phaseControl);
-            if (Math.abs(pValue - phaseControl.getTargetValue()) > halfTargetDeadband) {
-                var controllerContext = contextData.getControllersContexts().get(controllerBranch.getId());
-                double dp = phaseControl.getTargetValue() - pValue;
-                double a2p = sensitivityContext.calculateSensitivityFromA2P(controllerBranch, controlledBranch, phaseControl.getControlledSide());
-                if (Math.abs(a2p) > SENSI_EPS) {
-                    double da = Math.toRadians(dp / a2p);
-                    LOGGER.trace("Controlled branch '{}' active power is {} MW and out of target value {} MW (half deadband={} MW), a phase shift of {}° is required",
-                            controlledBranch.getId(), pValue * PerUnit.SB, phaseControl.getTargetValue() * PerUnit.SB, halfTargetDeadband * PerUnit.SB, Math.toDegrees(da));
-                    PiModel piModel = controllerBranch.getPiModel();
-
-                    int oldTapPosition = piModel.getTapPosition();
-                    Range<Integer> tapPositionRange = piModel.getTapPositionRange();
-                    piModel.updateTapPositionToReachNewA1(da, MAX_TAP_SHIFT, controllerContext.getAllowedDirection()).ifPresent(direction -> {
-                        controllerContext.updateAllowedDirection(direction);
-                        updated.setValue(true);
-                    });
-
-                    if (piModel.getTapPosition() != oldTapPosition) {
-                        LOGGER.debug("Controller branch '{}' change tap from {} to {} to reach active power target (full range: {})", controllerBranch.getId(),
-                                oldTapPosition, piModel.getTapPosition(), tapPositionRange);
-                    }
-                }
-            }
-        }
-
-        return updated.booleanValue();
     }
 
     @Override
