+++ src/main/java/com/powsybl/openloadflow/lf/outerloop/AbstractIncrementalPhaseControlOuterLoop.java
@@ -6,13 +6,21 @@
  */
 package com.powsybl.openloadflow.lf.outerloop;
 
+import com.powsybl.math.matrix.DenseMatrix;
+import com.powsybl.openloadflow.equations.EquationSystem;
+import com.powsybl.openloadflow.equations.EquationTerm;
+import com.powsybl.openloadflow.equations.JacobianMatrix;
 import com.powsybl.openloadflow.equations.Quantity;
 import com.powsybl.openloadflow.lf.AbstractLoadFlowParameters;
 import com.powsybl.openloadflow.lf.LoadFlowContext;
 import com.powsybl.openloadflow.network.*;
 import com.powsybl.openloadflow.util.PerUnit;
+import org.apache.commons.lang3.Range;
+import org.apache.commons.lang3.mutable.MutableBoolean;
+import org.slf4j.Logger;
 
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -29,6 +37,7 @@ public abstract class AbstractIncrementalPhaseControlOuterLoop<V extends Enum<V>
     public static final double MIN_TARGET_DEADBAND = 1 / PerUnit.SB; // 1 MW
     public static final double SENSI_EPS = 1e-6;
     public static final double PHASE_SHIFT_CROSS_IMPACT_MARGIN = 0.75;
+    public static Logger LOGGER;
 
     @Override
     public void initialize(O context) {
@@ -45,6 +54,91 @@ public abstract class AbstractIncrementalPhaseControlOuterLoop<V extends Enum<V>
 
     public static double getHalfTargetDeadband(TransformerPhaseControl phaseControl) {
         return Math.max(phaseControl.getTargetDeadband(), MIN_TARGET_DEADBAND) / 2;
+    }
+
+    public abstract static class AbstractSensitivityContext<V extends Enum<V> & Quantity, E extends Enum<E> & Quantity> {
+
+        private final List<LfBranch> controllerBranches;
+        private final EquationSystem<V, E> equationSystem;
+        private final JacobianMatrix<V, E> jacobianMatrix;
+        private final int[] controllerBranchIndex;
+        private DenseMatrix sensitivities;
+
+        public AbstractSensitivityContext(LfNetwork network, List<LfBranch> controllerBranches,
+                                  EquationSystem<V, E> equationSystem,
+                                  JacobianMatrix<V, E> jacobianMatrix) {
+            this.controllerBranches = Objects.requireNonNull(controllerBranches);
+            this.equationSystem = Objects.requireNonNull(equationSystem);
+            this.jacobianMatrix = Objects.requireNonNull(jacobianMatrix);
+            controllerBranchIndex = LfBranch.createIndex(network, controllerBranches);
+        }
+
+        private DenseMatrix getSensitivities() {
+            if (sensitivities == null) {
+                sensitivities = calculateSensitivityValues(controllerBranches, controllerBranchIndex, equationSystem, jacobianMatrix);
+            }
+            return sensitivities;
+        }
+
+        public abstract DenseMatrix calculateSensitivityValues(List<LfBranch> controllerBranches, int[] controllerBranchIndex,
+                                                               EquationSystem<V, E> equationSystem,
+                                                               JacobianMatrix<V, E> jacobianMatrix);
+
+        private EquationTerm<V, E> getP1(LfBranch controlledBranch) {
+            return (EquationTerm<V, E>) controlledBranch.getP1();
+        }
+
+        private EquationTerm<V, E> getP2(LfBranch controlledBranch) {
+            return (EquationTerm<V, E>) controlledBranch.getP2();
+        }
+
+        protected double calculateSensitivityFromA2S(LfBranch controllerBranch, EquationTerm<V, E> s) {
+            return s.calculateSensi(getSensitivities(), controllerBranchIndex[controllerBranch.getNum()]);
+        }
+
+        public double calculateSensitivityFromA2P(LfBranch controllerBranch, LfBranch controlledBranch, ControlledSide controlledSide) {
+            var p = controlledSide == ControlledSide.ONE ? getP1(controlledBranch) : getP2(controlledBranch);
+            return calculateSensitivityFromA2S(controllerBranch, p);
+        }
+    }
+
+    protected static boolean checkActivePowerControlPhaseControls(AbstractSensitivityContext sensitivityContext, IncrementalContextData contextData,
+                                                                  List<TransformerPhaseControl> activePowerControlPhaseControls) {
+        MutableBoolean updated = new MutableBoolean(false);
+
+        for (TransformerPhaseControl phaseControl : activePowerControlPhaseControls) {
+            LfBranch controllerBranch = phaseControl.getControllerBranch();
+            LfBranch controlledBranch = phaseControl.getControlledBranch();
+            var p = phaseControl.getControlledSide() == ControlledSide.ONE
+                    ? controlledBranch.getP1() : controlledBranch.getP2();
+            double pValue = p.eval();
+            double halfTargetDeadband = getHalfTargetDeadband(phaseControl);
+            if (Math.abs(pValue - phaseControl.getTargetValue()) > halfTargetDeadband) {
+                var controllerContext = contextData.getControllersContexts().get(controllerBranch.getId());
+                double dp = phaseControl.getTargetValue() - pValue;
+                double a2p = sensitivityContext.calculateSensitivityFromA2P(controllerBranch, controlledBranch, phaseControl.getControlledSide());
+                if (Math.abs(a2p) > SENSI_EPS) {
+                    double da = Math.toRadians(dp / a2p);
+                    LOGGER.trace("Controlled branch '{}' active power is {} MW and out of target value {} MW (half deadband={} MW), a phase shift of {}° is required",
+                            controlledBranch.getId(), pValue * PerUnit.SB, phaseControl.getTargetValue() * PerUnit.SB, halfTargetDeadband * PerUnit.SB, Math.toDegrees(da));
+                    PiModel piModel = controllerBranch.getPiModel();
+
+                    int oldTapPosition = piModel.getTapPosition();
+                    Range<Integer> tapPositionRange = piModel.getTapPositionRange();
+                    piModel.updateTapPositionToReachNewA1(da, MAX_TAP_SHIFT, controllerContext.getAllowedDirection()).ifPresent(direction -> {
+                        controllerContext.updateAllowedDirection(direction);
+                        updated.setValue(true);
+                    });
+
+                    if (piModel.getTapPosition() != oldTapPosition) {
+                        LOGGER.debug("Controller branch '{}' change tap from {} to {} to reach active power target (full range: {})", controllerBranch.getId(),
+                                oldTapPosition, piModel.getTapPosition(), tapPositionRange);
+                    }
+                }
+            }
+        }
+
+        return updated.booleanValue();
     }
 
 }
